201: Resolve comments from PR #250.
