+++ sdk/tests/ClientTest.cc
@@ -116,48 +116,3 @@ TEST_F(ClientTest, SetDefaultMaxTransactionFee)
   // Negative value should throw
   EXPECT_THROW(client.setMaxTransactionFee(fee.negated()), std::invalid_argument);
 }
-
-//-----
-TEST_F(ClientTest, ConnectToLocalNode)
-{
-  // Given
-  const std::string_view testAccountIdStr = "0.0.3";
-  const std::string_view jsonNetworkTag = "network";
-  const std::string_view jsonOperatorIdTag = "0.0.2";
-  const std::string_view jsonOperatorPrivateKey =
-    "302e020100300506032b65700422042091132178e72057a1d7528025956fe39b0b847f200ab59b2fdd367017f3087137";
-  const std::string testPathToJSON = getPathToJSON();
-  const std::unique_ptr<PrivateKey> testPrivateKey = ED25519PrivateKey::generatePrivateKey();
-  const std::shared_ptr<PublicKey> testPublicKey = testPrivateKey->getPublicKey();
-  const auto testInitialHbarBalance = Hbar(1000ULL, HbarUnit::TINYBAR());
-
-  std::ifstream testInputFile(testPathToJSON, std::ios::in);
-  std::string nodeAddress;
-  json jsonData;
-
-  EXPECT_NO_THROW(jsonData = json::parse(testInputFile));
-
-  if (jsonData[jsonNetworkTag][testAccountIdStr].is_string())
-  {
-    nodeAddress = jsonData[jsonNetworkTag][testAccountIdStr];
-  }
-
-  testInputFile.close();
-
-  std::unordered_map<std::string, AccountId> networkMap;
-  networkMap.insert(std::pair<std::string, AccountId>(nodeAddress, AccountId::fromString(testAccountIdStr)));
-
-  // When
-  Client client = Client::forNetwork(networkMap);
-  client.setOperator(AccountId::fromString(jsonOperatorIdTag), ED25519PrivateKey::fromString(jsonOperatorPrivateKey));
-
-  TransactionResponse txResp =
-    AccountCreateTransaction().setKey(testPublicKey).setInitialBalance(testInitialHbarBalance).execute(client);
-
-  const AccountId newAccountId = txResp.getReceipt(client).getAccountId().value();
-
-  // Then
-  EXPECT_EQ(client.getOperatorAccountId()->toString(), jsonOperatorIdTag);
-  EXPECT_NE(client.getOperatorPublicKey(), nullptr);
-  EXPECT_FALSE(newAccountId.toString().empty());
-}
