+++ sdk/tests/integration/CMakeLists.txt
@@ -1,0 +1,24 @@
+if (APPLE OR WIN32)
+    include(FetchContent)
+    FetchContent_Declare(googletest
+            GIT_REPOSITORY https://github.com/google/googletest.git
+            GIT_TAG release-1.12.1)
+
+    set(INSTALL_GTEST OFF CACHE BOOL "" FORCE)
+    FetchContent_MakeAvailable(googletest)
+    include(googletest)
+else ()
+    find_package(GTest REQUIRED)
+endif ()
+
+
+set(TEST_PROJECT_NAME ${PROJECT_NAME}-integration-tests)
+add_executable(${TEST_PROJECT_NAME}
+        ClientIntegrationTest.cc)
+
+target_link_libraries(${TEST_PROJECT_NAME} PRIVATE gtest_main gtest ${PROJECT_NAME})
+target_link_libraries(${TEST_PROJECT_NAME} PRIVATE nlohmann_json::nlohmann_json)
+gtest_discover_tests(${TEST_PROJECT_NAME} WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_BUILD_TYPE})
+
+file(COPY ${PROJECT_SOURCE_DIR}/config/local_node.json
+        DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_BUILD_TYPE})
